Create 3. Annuity payment
Calculations based on the formula:

A = P *(((i * (1 + i ) ** n) )/ (((1 + i) ** n) - 1))
where:
A - annuity payment.
P - loan principal.
i - nominal (monthly) interest rate. Usually, it’s 1/12 of the annual interest rate, and usually, it’s a floating value, not a percentage.
n - number of payments. This is usually the number of months in which repayments will be made.